Use big int for user id

+++ roller_bot/models/bonus.py
@@ -1,5 +1,5 @@
 
-from sqlalchemy import Column, DateTime, ForeignKey, Integer
+from sqlalchemy import Column, DateTime, ForeignKey, Integer, BigInteger
 from sqlalchemy.orm import relationship
 from roller_bot.models.base import Base
 
@@ -10,7 +10,7 @@ class Bonus(Base):
         Integer, primary_key=True,
         autoincrement=True, nullable=False
         )
-    user_id: Column = Column(Integer, ForeignKey('users.id'))
+    user_id: Column = Column(BigInteger, ForeignKey('users.id'))
     item_id: Column = Column(Integer, nullable=False)
     bonus_value: Column = Column(Integer, nullable=False, default=0)
     started_at: Column = Column(DateTime, nullable=False)
