+++ roller_bot/models/bonus_value.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, ForeignKey, Integer
+from sqlalchemy import Column, DateTime, ForeignKey, Integer, BigInteger
 from sqlalchemy.orm import relationship
 from roller_bot.models.base import Base
 
@@ -9,7 +9,7 @@ class BonusValue(Base):
             Integer, primary_key=True,
             autoincrement=True, nullable=False
     )
-    user_id: Column = Column(Integer, ForeignKey('users.id'), nullable=False)
+    user_id: Column = Column(BigInteger, ForeignKey('users.id'), nullable=False)
     roll_id: Column = Column(Integer, ForeignKey('rolls.id'), nullable=False)
     item_id: Column = Column(Integer, nullable=False)
     value: Column = Column(Integer, nullable=False, default=0)
