+++ roller_bot/models/items.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from sqlalchemy import Column, DateTime, ForeignKey, Integer
+from sqlalchemy import Column, DateTime, ForeignKey, Integer, BigInteger
 from sqlalchemy.orm import relationship
 
 from roller_bot.models.base import Base
@@ -13,7 +13,7 @@ class Items(Base):
             Integer, primary_key=True,
             autoincrement=True, nullable=False
     )
-    user_id: Column = Column(Integer, ForeignKey('users.id'), nullable=False)
+    user_id: Column = Column(BigInteger, ForeignKey('users.id'), nullable=False)
     item_id: Column = Column(Integer, nullable=False)
     quantity: Column = Column(Integer, nullable=False, default=1)
     health: Column = Column(Integer, nullable=False, default=100)
